+++ godot/code/Resets.gd
@@ -1,0 +1,4 @@
+extends Node
+
+func reset_networking_code():
+    NetworkingCode.players_data = {}