+++ godot/scenes/Lobbies/NetworkLobby.gd
@@ -1,12 +1,10 @@
 extends Control
 
-onready var ReadyGameBtn = get_node("ReadyGameBtn")
 onready var ConnectPopup = get_node("ConnectToHostPopup")
-onready var HostIP = get_node("ColorRect3/HostIP")
 
 func _ready():
     add_to_group("lobby")
-    if SettingsConf.S.QuickStart.host:
+    if SetConf.Session.host:
         pass
         # TODO: Change to the correct lobby type.
     else:
@@ -17,16 +15,14 @@ func first_frame_show_connect_popup():
     ConnectPopup.popup()
 
 func _on_BackToSetup_pressed():
-    SettingsConf.save()
-    if SettingsConf.S.QuickStart.host:
+    if SetConf.Session.host:
         SceneManager.goto_scene("res://scenes/Setups/NetworkSetup1.tscn")
     else:
         SceneManager.goto_scene("res://scenes/Setups/NetworkSetup0.tscn")
 
 # DEAD Code Below
 func _on_ReadyGame_pressed():
-    SettingsConf.S.QuickStart.quick_start_complete = true
-    SettingsConf.save()
+    SetConf.Session.quick_start_complete = true
     if not LazerInterface.state_lazer_gun_is_connected:
         get_tree().call_group("connect_weapon", "connect_weapon_guard", "res://scenes/InGame/InGameNetwork.tscn")
     else:
@@ -37,5 +33,5 @@ func _on_ReadyGame_pressed():
 ####################################
 func lobby_host_ip(Ip):
     ConnectPopup.hide()
-    NetworkingCode.server_ip = Ip
+    SetConf.Session.server_ip = Ip
     NetworkingCode.setup_as_client()