+++ godot/scenes/MainMenu/MainMenu.gd
@@ -8,7 +8,7 @@ onready var NonBlockingPopup = get_node("NonBlockingPopup")
 # Called when the node enters the scene tree for the first time.
 func _ready():
     add_to_group("lobby")
-    if not SettingsConf.S.QuickStart.quick_start_complete:
+    if not SetConf.Session.quick_start_complete:
         QuickStartBtn.disabled = true
     call_deferred("first_frame")
         
@@ -17,9 +17,11 @@ func _ready():
 #    pass
 
 func first_frame():
-    if SettingsConf.S.QuickStart.player_name == null:
+    if SetConf.Session.player_name == "":
         NonBlockingPopup.add_content("res://scenes/Popups/MissingPlayerNamePopup.tscn")
         NonBlockingPopup.popup()
+    var temp = load("res://code/Resets.gd").new()
+    temp.reset_networking_code()
 
 func _on_NoNetwork_pressed():
     SceneManager.goto_scene("res://scenes/Setups/NoNetworkSetup.tscn")
@@ -30,9 +32,6 @@ func _on_Testing_pressed():
 
 
 func _on_QuickStart_pressed():
-    SettingsConf.S.QuickStart.player_team
-    SettingsConf.S.QuickStart.player_number
-    SettingsConf.S.QuickStart.player_id
     if not LazerInterface.state_lazer_gun_is_connected:
         get_tree().call_group("connect_weapon", "connect_weapon_guard", "res://scenes/InGame/InGameNoNetwork.tscn")
     else: