+++ godot/scenes/SetupSections/TeamSetupSection.gd
@@ -8,8 +8,8 @@ onready var NumOfTeams = get_node("NumOfTeams")
 onready var HowManyLbl = get_node("HowManyTeamsLbl")
 # Called when the node enters the scene tree for the first time.
 func _ready():
-    TeamsBtn.pressed = SettingsConf.S.QuickStart.teams
-    NumOfTeams.set_val(SettingsConf.S.QuickStart.num_of_teams)
+    TeamsBtn.pressed = SetConf.Session.teams
+    NumOfTeams.set_val(SetConf.Session.num_of_teams)
     _on_TeamsBtn_pressed()
 
 func teams_active():
@@ -21,7 +21,7 @@ func teams_deactivated():
     NumOfTeams.visible = false
 
 func _on_TeamsBtn_pressed():
-    SettingsConf.S.QuickStart.teams = TeamsBtn.pressed
+    SetConf.Session.teams = TeamsBtn.pressed
     if TeamsBtn.pressed:
         NumOfTeams.set_val(2)
         teams_active()
@@ -32,13 +32,13 @@ func _on_TeamsBtn_pressed():
     
 
 func _on_NumOfTeams_PMWidChanged(new_val):
-    if SettingsConf.S.QuickStart.teams:
+    if SetConf.Session.teams:
         if new_val <= 1:
             NumOfTeams.set_val(2)
         elif new_val >= LazerInterface.MAX_TEAMS + 1:
             NumOfTeams.set_val(LazerInterface.MAX_TEAMS)
         else:
-            SettingsConf.S.QuickStart.num_of_teams = new_val
+            SetConf.Session.num_of_teams = new_val
     else:
-        SettingsConf.S.QuickStart.num_of_teams = new_val
-    emit_signal("on_num_of_teams_changed", SettingsConf.S.QuickStart.num_of_teams)
+        SetConf.Session.num_of_teams = new_val
+    emit_signal("on_num_of_teams_changed", SetConf.Session.num_of_teams)