+++ godot/scenes/Setups/GoToLobby.gd
@@ -13,5 +13,4 @@ func _ready():
 #    pass
 
 func _on_GoToLobby_pressed():
-    SettingsConf.save()
     SceneManager.goto_scene("res://scenes/Lobbies/NoNetworkLobby.tscn")