+++ godot/scenes/Setups/NetworkSetupCustom2.gd
@@ -1,6 +1,6 @@
 extends Control
 
-var ColorPick = preload("res://scenes/Widgets/TeamColorPicker.tscn")
+var ColorPick = load("res://scenes/Widgets/TeamColorPicker.tscn") 
 
 onready var TeamsColors = get_node("ColorRect/ScrollContainer/TeamsColorsContainer")
 
@@ -15,29 +15,30 @@ func _ready():
 func on_teams_changed():
     for i in TeamsColors.get_children():
         i.queue_free()
-    if SettingsConf.S.QuickStart.teams:
-        for team in range(1, SettingsConf.S.QuickStart.num_of_teams + 1):
+    if SetConf.Session.teams:
+        for team in range(1, SetConf.Session.num_of_teams + 1):
             var temp = ColorPick.instance()
             temp.text = "Team #" + str(team) + "'s Color ="
             temp.team_num = team
             TeamsColors.add_child(temp)
             #yield(get_tree(), "idle_frame")
-            if SettingsConf.S.QuickStart.TeamColors.size() >= team:
-                temp.set_color(SettingsConf.S.QuickStart.TeamColors[team - 1])
+            if SetConf.Session.team_colors.size() >= team:
+                temp.set_color(SetConf.Session.team_colors[team - 1])
             else:
-                SettingsConf.S.QuickStart.TeamColors.append("ffffff")
+                SetConf.Session.team_colors.append("ffffff")
                 temp.set_color("ffffff")
 
 
 func _on_TeamSetupSection_on_teams_changed(teams_active):
+    teams_active = null
     on_teams_changed()
 
 
 func _on_TeamSetupSection_on_num_of_teams_changed(num_of_teams):
+    num_of_teams = null
     on_teams_changed()
 
 
 func _on_ContinueSetup_pressed():
-    SettingsConf.save()
     NetworkingCode.set_lobby_team_or_ffa_as_server(null)
     SceneManager.goto_scene("res://scenes/Setups/NetworkSetupCustom3.tscn")